+++ src/MauiControlsExtras/Controls/ComboBox.xaml.cs
@@ -4,6 +4,7 @@ using System.Windows.Input;
 using MauiControlsExtras.Base;
 using MauiControlsExtras.Base.Validation;
 using MauiControlsExtras.Converters;
+using MauiControlsExtras.Helpers;
 using Microsoft.Maui.Controls.Shapes;
 
 namespace MauiControlsExtras.Controls;
@@ -804,6 +805,8 @@ public partial class ComboBox : TextStyledControlBase, IValidatable, Base.IKeybo
             textBox.PreviewKeyDown += OnWindowsTextBoxPreviewKeyDown;
         }
 #endif
+
+        MobileClipboardBridge.Setup(searchEntry, this);
     }
 
     private void OnKeyboardCaptureEntryHandlerChanged(object? sender, EventArgs e)